+++ src/ts/components/extensions/richtexteditor/RichTextEditor.tsx
@@ -179,3 +179,4 @@ const RichTextEditor = (props: Props) => {
 };
 
 export default RichTextEditor;
+