+++ src/ts/index.ts
@@ -177,8 +177,10 @@ import NavigationProgressProvider from './components/extensions/nprogress/Naviga
 import MantineProvider from './components/styles/MantineProvider';
 import ChipGroup from './components/core/chip/ChipGroup';
 import RichTextEditor from './components/extensions/richtexteditor/RichTextEditor';
+import { getEditor } from './utils/editorRegistry';
 import TypographyStylesProvider from './components/styles/TypographyStylesProvider';
 import DirectionProvider from './components/styles/DirectionProvider';
+
 
 export {
     Accordion,
@@ -358,6 +360,7 @@ export {
     VisuallyHidden,
     YearPickerInput,
     RichTextEditor,
+    getEditor,
     TypographyStylesProvider,
     appNotifications,
     DirectionProvider,